♻️(libvideo) move stats query to lib_video
Move useStatsVideo from lti queries to lib_video
