+++ src/frontend/apps/lti_site/data/queries/index.spec.tsx
@@ -21,7 +21,6 @@ import {
   useOrganization,
   usePlaylist,
   usePlaylists,
-  useStatsVideo,
   useThumbnail,
   useTimedTextTracks,
   useUpdatePlaylist,
@@ -668,54 +667,6 @@ describe('queries', () => {
           playlist: document.playlist.id,
           title: document.title,
         }),
-      });
-      expect(result.current.data).toEqual(undefined);
-      expect(result.current.status).toEqual('error');
-    });
-  });
-
-  describe('useStatsVideo', () => {
-    it('updates the resource', async () => {
-      const video = videoMockFactory();
-      fetchMock.get(`/api/videos/${video.id}/stats/`, {
-        nb_views: 123,
-      });
-
-      const { result, waitFor } = renderHook(() => useStatsVideo(video.id), {
-        wrapper: Wrapper,
-      });
-      await waitFor(() => result.current.isSuccess);
-
-      expect(fetchMock.lastCall()![0]).toEqual(
-        `/api/videos/${video.id}/stats/`,
-      );
-      expect(fetchMock.lastCall()![1]).toEqual({
-        headers: {
-          Authorization: 'Bearer some token',
-          'Content-Type': 'application/json',
-        },
-      });
-      expect(result.current.data).toEqual({ nb_views: 123 });
-      expect(result.current.status).toEqual('success');
-    });
-
-    it('fails to update the resource', async () => {
-      const video = videoMockFactory();
-      fetchMock.get(`/api/videos/${video.id}/stats/`, 400);
-
-      const { result, waitFor } = renderHook(() => useStatsVideo(video.id), {
-        wrapper: Wrapper,
-      });
-      await waitFor(() => result.current.isError);
-
-      expect(fetchMock.lastCall()![0]).toEqual(
-        `/api/videos/${video.id}/stats/`,
-      );
-      expect(fetchMock.lastCall()![1]).toEqual({
-        headers: {
-          Authorization: 'Bearer some token',
-          'Content-Type': 'application/json',
-        },
       });
       expect(result.current.data).toEqual(undefined);
       expect(result.current.status).toEqual('error');
